+++ osx/KA-Lite/KA-Lite/AppDelegate.m
@@ -17,12 +17,12 @@
 
 @implementation AppDelegate
 
-@synthesize startKalite, stopKalite, openInBrowserMenu, kaliteVersion, customKaliteData, startOnLogin, kaliteDataHelp, popover, popoverMsg;
+@synthesize startKalite, stopKalite, openInBrowserMenu, kaliteVersion, customKaliteData, loadOnLogin, startOnLoad, autoStartOnLoad, kaliteDataHelp, popover, popoverMsg, version, isLoaded;
 
 
 // REF: http://objcolumnist.com/2009/08/09/reopening-an-applications-main-window-by-clicking-the-dock-icon/
 - (BOOL)applicationShouldHandleReopen:(NSApplication *)theApplication hasVisibleWindows:(BOOL)flag {
-    if(flag==NO) {
+    if(flag == NO) {
         [self showPreferences];
     }
     return YES;	
@@ -35,18 +35,36 @@
 //}
 
 
+- (void)applicationWillFinishLaunching:(NSNotification *)aNotification {
+    // TODO(cpauya): Get version from the project's .plist file or from `kalite --version`.
+    self.version = @"0.16";
+    self.isLoaded = NO;
+    self.autoStartOnLoad = YES;
+    self.status = statusCouldNotDetermineStatus;
+    self.lastStatus = statusCouldNotDetermineStatus;
+    self.quitReason = quitByUnknown;
+}
+
+
 //<##>applicationDidFinishLaunching
 - (void)applicationDidFinishLaunching:(NSNotification *)aNotification {
 
+    // Set the delegate to self to make sure notifications work properly.
     [[NSUserNotificationCenter defaultUserNotificationCenter] setDelegate:self];
     
-    if (!checkKaliteExecutable()) {
-        NSLog(@"kalite executable is not found.");
-        [self showStatus:statusFailedToStart];
-        alert(@"Kalite executable is not found. You need to reinstall the KA Lite application.");
-        // The application must terminate if kalite executable is not found.
+    // MUST: Let's check the setup if everything is good!
+    if ([self checkSetup:YES] == NO) {
+        // The application must terminate if setup is not good.
+        void *sel = @selector(closeSplash);
+        alert(@"The KA Lite installation is not complete, please re-install KA Lite. \n\nRefer to the Console app for details.");
         [[NSApplication sharedApplication] terminate:nil];
+        return;
     }
+    
+    // Setup is good, let's continue.
+    
+    // Make sure to register default values for the user preferences.
+    [self registerDefaultPreferences];
     
     self.statusItem = [[NSStatusBar systemStatusBar] statusItemWithLength:NSVariableStatusItemLength];
     [self.statusItem setImage:[NSImage imageNamed:@"favicon"]];
@@ -56,49 +74,65 @@
     
     [self.kaliteDataHelp setToolTip:@"This will set the KALITE_HOME environment variable to the selected KA Lite data location. \n \nClick the 'Apply' button to save your changes and click the 'Start KA Lite' button to use your new data location. \n \nNOTE: To use your existing KA Lite data, manually copy it to the selected KA Lite data location."];
     [self.kaliteUninstallHelp setToolTip:@"This will uninstall the KA Lite application. \n \nCheck the `Delete KA Lite data folder` option if you want to delete your KA Lite data. \n \nNOTE: This will require admin privileges."];
-    
-    // Set the default status.
-    self.status = statusCouldNotDetermineStatus;
-    [self getKaliteStatus];
-    
-    @try {
-        checkEnvVars();
-        NSString *database = getDatabasePath();
-        NSString *kalite = getKaliteExecutable();
-        if (!pathExists(database)) {
-            NSLog(@"Database not found, must show preferences.");
-        } else {
-            NSLog([NSString stringWithFormat:@"FOUND database at %@!", database]);
-        }
-        NSLog([NSString stringWithFormat:@"FOUND kalite at %@!", kalite]);
-        showNotification(@"KA Lite is now loaded.");
-        [self runKalite:@"--version"];
-        [self getKaliteStatus];
-    }
-    @catch (NSException *ex) {
-        NSLog(@"KA Lite had an Error: %@", ex);
-    }
+
+//    @try {
+//        [self runKalite:@"--version"];
+//        [self getKaliteStatus];
+//    }
+//    @catch (NSException *ex) {
+//        NSLog(@"KA Lite had an Error: %@", ex);
+//    }
     
     void *sel = @selector(closeSplash);
     [NSTimer scheduledTimerWithTimeInterval:0.1 target:self selector:sel userInfo:nil repeats:NO];
     [self startKaliteTimer];
+
+    // TODO(cpauya): Auto-start KA Lite on application load.
+    if (self.autoStartOnLoad) {
+        [self startFunction];
+    } else {
+        // Get the status to determine the menu bar icon to display but don't show any notifications.
+        // The `isLoaded` property will be set to YES the initial status check.
+        showNotification(@"KA Lite is now loaded, click on the Start KA Lite menu to get started.", @"");
+        [self getKaliteStatus];
+    }
 
 }
 
 
 - (NSApplicationTerminateReply)applicationShouldTerminate:(NSApplication *)sender {
     // Confirm quit action from user.
-    NSString *msg = @"This will stop and quit KA Lite, are you sure?";
-    if (! confirm(msg)) {
-        return NSTerminateCancel;
+    // TODO(cpauya): Don't ask if OS asked to quit the app.
+    if ([self checkSetup:NO] == YES) {
+        NSString *msg = @"This will stop and quit KA Lite, are you sure?";
+        switch ([self quitReason]) {
+            case quitByUser:
+                break;
+            default:
+                if (! confirm(msg)) {
+                    return NSTerminateCancel;
+                }
+                break;
+        }
+        [self stopFunction:true];
     }
-    [self stopFunction:true];
     return NSTerminateNow;
 }
 
 
 - (void)applicationWillTerminate:(NSNotification *)aNotification {
-    showNotification(@"KA Lite is now stopped and quit.  Thank you.");
+    NSString *msg = @"KA Lite is now stopped and quit.";
+    if ([self checkSetup:NO] == NO) {
+        switch ([self quitReason]) {
+            case quitByUninstall:
+                msg = @"KA Lite was quit to complete the uninstall process.";
+                break;
+            default:
+                msg = @"KA Lite was quit because the required setup is incomplete.";
+                break;
+        }
+    }
+    showNotification(msg, @"");
 }
 
 
@@ -114,15 +148,16 @@
 
 
 BOOL checkEnvVars() {
-    // MUST: Check the KALITE_PYTHON environment variable
-    // and default it to the .app Resources folder if not yet set.
     NSString *kalitePython = getEnvVar(@"KALITE_PYTHON");
     if (!pathExists(kalitePython)) {
-        NSString *msg = [NSString stringWithFormat:@"The KALITE_PYTHON environment variable points to '%@' is not valid.", kalitePython];
-        showNotification(msg);
-        return FALSE;
+        return NO;
     }
-    return TRUE;
+    // If KALITE_HOME is nil or an empty string, that's fine.  Else check if value is a valid path.
+    NSString *kaliteHome = getEnvVar(@"KALITE_HOME");
+    if (kaliteHome && !pathExists(kaliteHome)) {
+        return NO;
+    }
+    return YES;
 }
 
 
@@ -201,9 +236,11 @@ BOOL checkEnvVars() {
     
     [[task.standardOutput fileHandleForReading] setReadabilityHandler:^(NSFileHandle *file) {
         NSData *data = [file availableData]; // this will read to EOF, so call only once
-        NSLog(@"%@", [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding]);
         NSString *outStr = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
-        [self displayLogs:outStr];
+        if (self.status != self.lastStatus) {
+            NSLog(@"%@", [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding]);
+            [self displayLogs:outStr];
+        }
         
         // Set the current kalite version
         if (command == versionStr){
@@ -298,7 +335,7 @@ BOOL kaliteExists() {
             if (status >= 255) {
                 status = status >> 8;
             }
-            self.status = status;
+            [self setNewStatus:status];
             if (oldStatus != status) {
                 [self showStatus:self.status];
             }
@@ -311,9 +348,9 @@ BOOL kaliteExists() {
             return self.status;
         }
     } else {
-        self.status = statusCouldNotDetermineStatus;
+        [self setNewStatus:statusCouldNotDetermineStatus];
         [self showStatus:self.status];
-        showNotification(@"The `kalite` executable does not exist!");
+        showNotification(@"The `kalite` executable does not exist!", @"");
     }
     return self.status;
 }
@@ -328,7 +365,7 @@ BOOL kaliteExists() {
         }
     }
     @catch (NSException *ex) {
-        self.status = statusCouldNotDetermineStatus;
+        [self setNewStatus:statusCouldNotDetermineStatus];
         NSLog(@"Error running `kalite` %@", ex);
     }
     return self.status;
@@ -355,25 +392,33 @@ BOOL confirm(NSString *message) {
 }
 
 
-void showNotification(NSString *subtitle) {
+void showNotification(NSString *subtitle, NSString *info) {
     // REF: http://stackoverflow.com/questions/12267357/nsusernotification-with-custom-soundname?rq=1
     // TODO(cpauya): These must be ticked by user on preferences if they want notifications, sounds, or not.
     NSUserNotification* notification = [[NSUserNotification alloc] init];
     notification.title = @"KA Lite";
     notification.subtitle = subtitle;
+    notification.informativeText = info;
     notification.soundName = @"Basso.aiff";
 
     NSUserNotificationCenter *nc = [NSUserNotificationCenter defaultUserNotificationCenter];
     [nc deliverNotification:notification];
     // The notification may be optional (based on user's OS X preferences) but we must show it on the logs.
     NSLog(subtitle);
+    if (info) {
+        NSLog(info);
+    }
 }
 
 
-- (void)disableKaliteDataPath{
-    // Disable custom kalite data path when kalite is still running.
-    self.customKaliteData.enabled = NO;
-    [self.customKaliteData setToolTip:@"KA Lite is still running. Stop KA Lite to select data path."];
+- (void)toggleKaliteDataPath:(BOOL)toggleValue {
+    if (toggleValue == YES) {
+        self.customKaliteData.enabled = YES;
+        [self.customKaliteData setToolTip:@"Select KA Lite data path."];
+    } else {
+        self.customKaliteData.enabled = NO;
+        [self.customKaliteData setToolTip:@"KA Lite is still running. Stop KA Lite to select data path."];
+    }
 }
 
 
@@ -447,6 +492,17 @@ NSString *getEnvVar(NSString *var) {
  ********************/
 
 
+/*
+ This will keep track of the last status so that we don't fill the Logs when we check for status every interval.
+ So instead of showing "running" status logs every minute, we just show the log when a status change is detected
+ based on the timer interval.
+ */
+- (void)setNewStatus:(enum kaliteStatus)newStatus {
+    self.lastStatus = self.status;
+    self.status = newStatus;
+}
+
+
 - (void)showStatus:(enum kaliteStatus)status {
     // Enable/disable menu items based on status.
     BOOL canStart = pathExists(getKaliteExecutable()) > 0 ? YES : NO;
@@ -460,11 +516,8 @@ NSString *getEnvVar(NSString *var) {
             [self.openInBrowserMenu setEnabled:NO];
             [self.statusItem setImage:[NSImage imageNamed:@"exclaim"]];
             [self.statusItem setToolTip:@"KA Lite failed to start."];
-            
             // Disable custom kalite data path when kalite is still running.
-            self.customKaliteData.enabled = NO;
-            [self.customKaliteData setToolTip:@"KA Lite failed to start"];
-            
+            [self toggleKaliteDataPath:NO];
             break;
         case statusStartingUp:
             [self.startKalite setEnabled:NO];
@@ -475,7 +528,8 @@ NSString *getEnvVar(NSString *var) {
             self.openBrowserButton.enabled = NO;
             [self.statusItem setToolTip:@"KA Lite is starting..."];
             [self.statusItem setImage:[NSImage imageNamed:@"loading"]];
-            [self disableKaliteDataPath];
+            // Disable custom kalite data path when kalite is still running.
+            [self toggleKaliteDataPath:NO];
             break;
         case statusOkRunning:
             [self.startKalite setEnabled:NO];
@@ -486,8 +540,9 @@ NSString *getEnvVar(NSString *var) {
             self.openBrowserButton.enabled = YES;
             [self.statusItem setImage:[NSImage imageNamed:@"stop"]];
             [self.statusItem setToolTip:@"KA Lite is running."];
-            showNotification(@"You can now click on 'Open in Browser' menu");
-            [self disableKaliteDataPath];
+            showNotification(@"Running, you can now click on 'Open in Browser' menu.", @"");
+            // Disable custom kalite data path when kalite is still running.
+            [self toggleKaliteDataPath:NO];
             break;
         case statusStopped:
             [self.startKalite setEnabled:canStart];
@@ -498,9 +553,15 @@ NSString *getEnvVar(NSString *var) {
             self.openBrowserButton.enabled = NO;
             [self.statusItem setImage:[NSImage imageNamed:@"favicon"]];
             [self.statusItem setToolTip:@"KA Lite is stopped."];
-            showNotification(@"Stopped");
-            self.customKaliteData.enabled = YES;
-            [self.customKaliteData setToolTip:@"Select KA Lite data path."];
+
+            // We don't want to show "Stopped" right after we have loaded the application and checked the status.
+            if (self.isLoaded) {
+                showNotification(@"Stopped", @"");
+            }
+
+            // Enable setting the custom kalite data path.
+            [self toggleKaliteDataPath:YES];
+            
             break;
         default:
             [self.startKalite setEnabled:canStart];
@@ -509,15 +570,14 @@ NSString *getEnvVar(NSString *var) {
             self.startButton.enabled = canStart;
             self.stopButton.enabled = NO;
             self.openBrowserButton.enabled = NO;
-            if (kaliteExists()){
+            if (kaliteExists()) {
                 [self.statusItem setImage:[NSImage imageNamed:@"favicon"]];
-            }else{
+            } else {
                 [self.statusItem setImage:[NSImage imageNamed:@"exclaim"]];
             }
-//            [self.statusItem setToolTip:@"KA-Lite has encountered an error, pls check the Console."];
-//            showNotification(@"Has encountered an error, pls check the Console.");
             break;
     }
+    self.isLoaded = YES;
 }
 
 
@@ -526,9 +586,8 @@ NSString *getEnvVar(NSString *var) {
         alert(@"KA Lite is still processing, please wait until it is finished.");
         return;
     }
-    showNotification(@"Starting...");
-    self.status = statusStartingUp;
-    [self showStatus:statusStartingUp];
+    showNotification(@"Starting...", @"");
+    [self setNewStatus:statusStartingUp];
     [self runKalite:@"start"];
 }
 
@@ -542,7 +601,7 @@ NSString *getEnvVar(NSString *var) {
     if (isQuit) {
         msg = @"Stopping and quitting the application...";
     }
-    showNotification(msg);
+    showNotification(msg, @"");
     [self runKalite:@"stop"];
 }
 
@@ -552,7 +611,7 @@ NSString *getEnvVar(NSString *var) {
     NSURL *url = [NSURL URLWithString:@"http://127.0.0.1:8008/"];
     if( ![[NSWorkspace sharedWorkspace] openURL:url] ) {
         NSString *msg = [NSString stringWithFormat:@" Failed to open url: %@",[url description]];
-        showNotification(msg);
+        showNotification(msg, @"");
     }
 }
 
@@ -582,7 +641,7 @@ NSString *getEnvVar(NSString *var) {
 }
 
 
-- (IBAction)startOnLogin:(id)sender {
+- (IBAction)loadOnLogin:(id)sender {
     self.savePrefs.enabled = TRUE;
 }
 
@@ -644,15 +703,17 @@ NSString *getEnvVar(NSString *var) {
             const char *runCommand = [[NSString stringWithFormat: @"%@ %@", kaliteUninstallPath, kaliteUninstallArg] UTF8String];
             int runCommandStatus = system(runCommand);
             if (runCommandStatus == 0) {
+                self.quitReason = quitByUninstall;
                 // Terminate application.
                 [[NSApplication sharedApplication] terminate:nil];
-            }else {
-                alert(@"The KA Lite uninstall did not succeed. You can see the logs at console application.");
+            } else {
+                alert(@"The KA Lite uninstall did not succeed. You can see the logs at the Console application.");
             }
         }
         
     } else {
-        alert(@"The KA Lite uninstall script is not found. You need to reinstall the KA Lite application.");
+        NSString *msg = [NSString stringWithFormat:@"The KA Lite uninstall script is not found at '%@'. You need to reinstall the KA Lite application.", kaliteUninstallPath];
+        alert(msg);
     }
 }
 
@@ -685,16 +746,100 @@ NSString *getEnvVar(NSString *var) {
 }
 
 
+/*
+ Checks if environment for running KA Lite is good:
+ 1. `kalite` executable exists
+ 2. environment variables: KALITE_PYTHON, KALITE_HOME
+ 3. Custom KA Lite data path.
+*/
+- (BOOL)checkSetup:(BOOL)showIt {
+    NSString *title = @"The KA Lite installation is incomplete.";
+    NSString *msg = @"";
+    BOOL isOk = YES;
+
+    // Check the kalite executable.
+    if (! checkKaliteExecutable()) {
+        msg = [NSString stringWithFormat:@"%@\n* The KA Lite executable cannot be found.", msg];
+        isOk = NO;
+    }
+
+    // Check the environment variables.
+    if (! checkEnvVars()) {
+        msg = [NSString stringWithFormat:@"%@\n* One of the KALITE_PYTHON or KALITE_HOME environment variables is invalid.", msg];
+        isOk = NO;
+    }
+
+    // Check the custom KA Lite data path.
+    NSString *dataPath = getKaliteDataPath();
+    if (dataPath == nil) {
+        msg = [NSString stringWithFormat:@"%@\n* The custom KA Lite data path is invalid, please check the KALITE_HOME environment variable value.", msg];
+        isOk = NO;
+    }
+
+    if (showIt == YES && isOk == NO) {
+        msg = [NSString stringWithFormat:@"%@  Please try to re-install KA Lite to attempt to fix the issue/s.%@", title, msg];
+        showNotification(title, msg);
+    }
+    return isOk;
+}
+
+
+// Checks for default preferences, sets them accordingly, and saves the .plist.
+// Returns YES if defaults preferences were set, otherwise NO.
+- (BOOL)registerDefaultPreferences {
+    
+    NSString *dataPath = getKaliteDataPath();
+    if (dataPath == nil) {
+        NSLog(@"The default KA Lite data path is nil, please check the KALITE_HOME environment variable value or re-install KA Lite.");
+        return NO;
+    }
+    NSDictionary *dict = @{
+                           @"version": self.version,
+                           @"autoLoadOnLogin": @YES,
+                           @"autoStartOnLoad": @YES,
+                           @"customKaliteData": dataPath
+                           };
+    
+    NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
+
+    // MUST: Check a key in the default preferences and override with the default values if non-extant.
+    NSNumber *version = [defaults objectForKey:@"version"];
+    if (version == nil) {
+        [defaults registerDefaults:dict];
+        [defaults setValuesForKeysWithDictionary:dict];
+        BOOL result = [defaults synchronize];
+        
+        // TODO(cpauya): Let's perform the actions based on the values of the preferences.
+        // 1. autoLoadOnLogin
+        // 2. autoStartOnLoad
+        [self setEnvVarsAndPlist];
+        return YES;
+    }
+    self.kaliteVersion.stringValue = version;
+    return NO;
+}
+
+
 - (void)loadPreferences {
+    
+    // MUST: Check for default preferences first before actually loading the stored preferences.
+    // It's possible that the .plist was deleted while the .app is loaded, so this makes sure
+    // we still load the default preferences just in case.
+    [self registerDefaultPreferences];
+    
     NSString *kaliteDataPath = getKaliteDataPath();
     if (!kaliteDataPath) {
-        showNotification(@"KA Lite data folder is not found. Click the `Start KA Lite` button to auto-create the KA Lite data folder.");
+        showNotification(@"KA Lite data folder is not found. Click the `Start KA Lite` button to auto-create the KA Lite data folder.", @"");
     }
     NSString *standardizedPath = [kaliteDataPath stringByStandardizingPath];
     self.customKaliteData.stringValue = standardizedPath;
     
-    if ([[NSUserDefaults standardUserDefaults] boolForKey:@"autoStartOnLogin"]){
-        self.startOnLogin.state = YES;
+    NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
+    NSNumber *state = [defaults objectForKey:@"autoLoadOnLogin"];
+    if (state == nil || [state boolValue]){
+        self.loadOnLogin.state = YES;
+    } else {
+        self.loadOnLogin.state = NO;
     }
 }
 
@@ -712,10 +857,12 @@ NSString *getEnvVar(NSString *var) {
     // REF: http:iosdevelopertips.com/core-services/encode-decode-using-base64.html
     NSUserDefaults *prefs = [NSUserDefaults standardUserDefaults];
     
-    //Set autoStartOnLogin value.
-    [prefs setBool:FALSE forKey:@"autoStartOnLogin"];
-    if ([self.startOnLogin state] == NSOnState) {
-        [prefs setBool:TRUE forKey:@"autoStartOnLogin"];
+    // Set autoLoadOnLogin value, defaults to YES.
+    NSInteger state = [self.loadOnLogin state];
+    if (state == NSOffState) {
+        [prefs setObject:[NSNumber numberWithBool:NO] forKey:@"autoLoadOnLogin"];
+    } else {
+        [prefs setObject:[NSNumber numberWithBool:YES] forKey:@"autoLoadOnLogin"];
     }
     
     NSString *customKaliteData = [[self.customKaliteData URL] path];
@@ -724,32 +871,17 @@ NSString *getEnvVar(NSString *var) {
     }
     
     // REF: https:github.com/iwasrobbed/Objective-C-CheatSheet#storing-values
-    [prefs synchronize];
-    
-    setEnvVarsAndPlist();
-    
-    // Automatically run `kalite manage setup` if no database was found.
-    NSString *databasePath = getDatabasePath();
-    if (!pathExists(databasePath)) {
-        if (checkKaliteExecutable()) {
-            alert(@"Will now run KA Lite setup, it will take a few minutes.  Please wait until prompted that setup is done.");
-            enum kaliteStatus status = [self setupKalite];
-            showNotification(@"Setup is finished!  You can now start KA Lite.");
-            [self.statusItem setImage:[NSImage imageNamed:@"exclaim"]];
-            // TODO(cpauya): Get the result of running `bin/kalite manage setup` not the
-            // default result of `bin/kalite status` so we can alert the user that setup failed.
-            //        if (status != statusStopped) {
-            //            alert(@"Running 'manage setup' failed, please see Console.");
-            //            return;
-            //        }
-        }
+    // Handle the NO Boolean return if unsuccessful.
+    BOOL result = [prefs synchronize];
+    if (! result) {
+        showNotification(@"Sorry but your preferences cannot be saved, please check the Console logs.", @"");
     }
     
-    // Close the preferences dialog after successful save.
-    [window orderOut:[window identifier]];
+    [self setEnvVarsAndPlist];
     
-    // Terminate application.
-//    [[NSApplication sharedApplication] terminate:nil];
+
+    // Close the preferences dialog after a successful save.
+    [window orderOut:[window identifier]];
 }
 
 
@@ -773,9 +905,10 @@ NSString *getEnvVar(NSString *var) {
 }
 
 
-BOOL setEnvVarsAndPlist() {
+- (BOOL)setEnvVarsAndPlist {
     /*
-    TODO(cpauya): This function does ...
+    This function sets the KALITE_HOME environment variable based on the custom KA Lite data path and
+    then it sets the .plist file contents so the env var is used when computer is rebooted.
     */
 
     // REF: http://stackoverflow.com/questions/99395/how-to-check-if-a-folder-exists-in-cocoa-objective-c
@@ -800,16 +933,16 @@ BOOL setEnvVarsAndPlist() {
     // Set the KALITE_HOME environment variable using the system() function
     // so that it will be updated if already set and used by the app.
     NSString *kaliteDataPath = getKaliteDataPath();
-    showNotification([NSString stringWithFormat:@"Setting KALITE_HOME environment variable to %@...", kaliteDataPath]);
+    showNotification([NSString stringWithFormat:@"Setting KALITE_HOME environment variable to %@...", kaliteDataPath], @"");
     if (!kaliteDataPath) {
-        showNotification(@"KA Lite data folder is not found. Click the `Start KA Lite` button to auto-create the KA Lite data folder.");
+        showNotification(@"KA Lite data folder is not found. Click the `Start KA Lite` button to auto-create the KA Lite data folder.", @"");
         return FALSE;
     }
     NSString *command = [NSString stringWithFormat:@"launchctl setenv KALITE_HOME \"%@\"", kaliteDataPath];
     const char *cmd = [command UTF8String];
     int i = system(cmd);
     if (i != 0) {
-        showNotification(@"Failed to set KALITE_HOME env.");
+        showNotification(@"Failed to set KALITE_HOME env.", @"");
         return FALSE;
     }
     
@@ -821,20 +954,24 @@ BOOL setEnvVarsAndPlist() {
     NSMutableDictionary *plistDict = [[NSMutableDictionary alloc] init];
     [plistDict setObject:plist forKey:@"Label"];
 
-    // If autoStartOnLogin value is TRUE, append the command to "open the Kalite.app" to the plist.
+    // If autoLoadOnLogin value is TRUE, append the command to "open the Kalite.app" to the plist.
     NSString *kaliteHomeStr = [NSString stringWithFormat:@"%@",
                                [NSString stringWithFormat:@"launchctl setenv KALITE_HOME \"%@\"", kaliteDataPath]
                                ];    
     NSString *launchStr = [NSString stringWithFormat:@"%@", kaliteHomeStr];
-    if ([[NSUserDefaults standardUserDefaults] boolForKey:@"autoStartOnLogin"]){
+    // Check for a not-NO here because the preference also be nil if not yet set, to which we treat it as YES since
+    // we want it to default to YES.
+    NSNumber *state = [[NSUserDefaults standardUserDefaults] objectForKey:@"autoLoadOnLogin"];
+    if ([state boolValue] != NO){
         NSString *kaliteAppPath = [NSString stringWithFormat:@"open %@", [[NSBundle mainBundle] bundlePath]];
-         launchStr = [NSString stringWithFormat:@"%@ ; %@", kaliteHomeStr, kaliteAppPath];
+        launchStr = [NSString stringWithFormat:@"%@ ; %@", kaliteHomeStr, kaliteAppPath];
     }
 
     // More contents for the .plist command.
     NSArray *arr = @[@"sh", @"-c", launchStr];
     [plistDict setObject:arr forKey:@"ProgramArguments"];
     [plistDict setObject:[NSNumber numberWithBool:TRUE] forKey:@"RunAtLoad"];
+    [plistDict setObject:self.version forKey:@"version"];
 
     // Write the formed content to set the KALITE_HOME env var to the .plist.
     NSLog([NSString stringWithFormat:@"Writing the .plist for the KALITE_HOME environment variable... %@", plistDict]);
@@ -846,18 +983,8 @@ BOOL setEnvVarsAndPlist() {
     NSLog([NSString stringWithFormat:@"Saved .plist file to %@", target]);
 
     NSString *msg = [NSString stringWithFormat:@"Successfully set KALITE_HOME env to %@.", kaliteDataPath];
-    showNotification(msg);
+    showNotification(msg, @"");
     return TRUE;
-}
-
-
--(enum kaliteStatus)setupKalite {
-    NSString *cmd = [NSString stringWithFormat:@"manage setup --noinput"];
-    NSString *msg = [NSString stringWithFormat:@"Running `kalite manage setup`"];
-    showNotification(msg);
-    enum kaliteStatus status = [self runKalite:cmd];
-    [self getKaliteStatus];
-    return status;
 }
 
 
@@ -870,7 +997,7 @@ BOOL setEnvVarsAndPlist() {
 - (void)startKaliteTimer {
     // TODO(cpauya): Use initWithFireDate of NSTimer instance.
     // TODO(amodia): Check if kalite environment variables change.
-    [NSTimer scheduledTimerWithTimeInterval:60.0
+    [NSTimer scheduledTimerWithTimeInterval:5.0
                                      target:self
                                    selector:@selector(getKaliteStatus)
                                    userInfo:nil